Working Mobile Site
The mobile page in this site was routed to and given there was no mobile page, it delivered a 404. This basic mobile site gives a splash landing page
to allow for indexing for mobile devices. Currently, the site is a stripped down version of the landing page, and only rudimentarily responsive.
In progress is full responsiveness and optimization based on the device in question and the dimensions and capabilities.